% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_indep_interaction.R
\name{ci.indep.interaction}
\alias{ci.indep.interaction}
\title{Confidence interval for the interaction of independent proportions}
\usage{
ci.indep.interaction(ci = NULL, a1 = NULL, m1 = NULL, b1 = NULL,
  n1 = NULL, a2 = NULL, m2 = NULL, b2 = NULL, n2 = NULL, p1 = NULL,
  l1 = NULL, u1 = NULL, p2 = NULL, l2 = NULL, u2 = NULL, p3 = NULL,
  l3 = NULL, u3 = NULL, p4 = NULL, l4 = NULL, u4 = NULL)
}
\arguments{
\item{ci}{The confidence level - pass as a natural number (e.g. 95
for the 95\% confidence interval). Only pass this argument when
passing absolute frequencies (a, m , b, n); do not combine with
proportions and confidence interval boundaries (in this case the
confidence interval is implied in the boundaries from the passed
intervals).}

\item{a1}{Frequency of events in proportion A1}

\item{m1}{Total frequency in proportion A1}

\item{b1}{Frequency of events in proportion B1}

\item{n1}{Total frequency in proportion B1}

\item{a2}{Frequency of events in proportion A2}

\item{m2}{Total frequency in proportion A2}

\item{b2}{Frequency of events in proportion B2}

\item{n2}{Total frequency in proportion B2}

\item{p1}{Proportion 1 - a value between 0 and 1. Pass only if
absolute frequencies (a, m , b, n) are not given.}

\item{l1}{The lower bound of the confidence interval of p1.}

\item{u1}{The upper bound of the confidence interval of p1.}

\item{p2}{Proportion 2 - a value between 0 and 1. Pass only if
absolute frequencies (a, m , b, n) are not given.}

\item{l2}{The lower bound of the confidence interval of p2.}

\item{u2}{The upper bound of the confidence interval of p2.}

\item{p3}{Proportion 3 - a value between 0 and 1. Pass only if
absolute frequencies (a, m , b, n) are not given.}

\item{l3}{The lower bound of the confidence interval of p3.}

\item{u3}{The upper bound of the confidence interval of p3.}

\item{p4}{Proportion 4 - a value between 0 and 1. Pass only if
absolute frequencies (a, m , b, n) are not given.}

\item{l4}{The lower bound of the confidence interval of p4.}

\item{u4}{The upper bound of the confidence interval of p4.}
}
\value{
A \code{list} containing the confidence interval boudaries.
  \item{d}{The difference between differences (d = (a1/m1 - b1/n1) - (a2/m2 - b2/n2) = (p1 - p2) - (p3 - p4)}
  \item{l}{The lower bound of the confidence interval}
  \item{u}{The upper bound of the confidence interval}
}
\description{
This function tests for an interaction effect in independent
proportions: it tests the whether the difference between two
proportions is larger than the difference between two different
proportions using a confidence interval (i.e., the parameter of
interest is d = (p1 - p2) - (p3 - p4)). This corresponds closely to a
2x2 ANOVA when outcomes are dichotomous rather than continuous. See
Newcombe (2001).
}
\examples{

# Example from From Newcombe (2001):

ci.indep.interaction(ci=95, a1=17, m1=65, b1=17, n1=75, a2=18, m2=72,
b2=16, n2=65)

# Alternative usage:
p1 <- ci.one.prop(95, 17, 65)
p2 <- ci.one.prop(95, 17, 75)
p3 <- ci.one.prop(95, 18, 72)
p4 <- ci.one.prop(95, 16, 65) # it is possible to use other CI boundaries
ci.indep.interaction(p1=p1$p, p2=p2$p, p3=p3$p, p4=p4$p, l1=p1$l, l2=p2$l, l3=p3$l, l4=p4$l,
                     u1=p1$u, u2=p2$u, u3=p3$u, u4=p4$u)

}
\references{
Newcombe, R. G. (2001). Estimating the difference between
      differences: measurement of additive scale interaction for
      proportions. Statistics in medicine, 20(19), 2885-2893.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
