% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mixed_between.R
\name{ci.mixed.between}
\alias{ci.mixed.between}
\title{Confidence interval for the "main effect" of a between-subjects
factor in mixed 2 x 2 design with a dichotomous outcome}
\usage{
ci.mixed.between(ci, group1.measure1, group1.measure2, group2.measure1,
  group2.measure2)
}
\arguments{
\item{ci}{The confidence level - pass as a natural number (e.g. 95
for the 95\% confidence interval).}

\item{group1.measure1}{A vector of "successes" (= 1) and "fails" (=
0) for the first measure in group 1}

\item{group1.measure2}{A vector of "successes" (= 1) and "fails" (=
0) for the second measure in group 1}

\item{group2.measure1}{A vector of "successes" (= 1) and "fails" (=
0) for the first measure in group 2}

\item{group2.measure2}{A vector of "successes" (= 1) and "fails" (=
0) for the second measure in group 2}
}
\description{
This function computes a confidence interval for the main effect of
a grouping variable across two levels of a within subject factor
when the measure of interest is dichotomous. This for example
corresponds to the effectiveness of a treatment on individuals
across two different points in time. The confidence interval that
is produced is corrected for the within-subjects dependence of the
two measures by applying a correction suggested by Donner and Klar
(section 2.1; 1993) to the standard error of the difference of
proportions.
}
\references{
Donner, A., & Klar, N. (1993). Confidence interval construction for
  effect measures arising from cluster randomization
  trials. Journal of clinical epidemiology, 46(2), 123-131.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
