% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_one_prop_cluster.R
\name{ci.one.prop.cluster}
\alias{ci.one.prop.cluster}
\title{Confidence interval for a single proportion estimated from a
clustered sample}
\usage{
ci.one.prop.cluster(ci, successes, N)
}
\arguments{
\item{ci}{The confidence level - pass as a natural number (e.g. 95
for the 95\% confidence interval)}

\item{successes}{A vector describing the number of "successes" (i.e.,
response = 1) in each cluster.}

\item{N}{A vector describing the total number of responses in each
cluster.}
}
\value{
A \code{list} containing the confidence interval boudaries.
  \item{p}{The estimated proportion of "success" across all clusters}
  \item{l}{The lower bound of the confidence interval of p}
  \item{u}{The upper bound of the confidence interval of p}
  \item{se.cluster}{The estimated standard error of the proportion}
  \item{design.effect}{The design effect}
  \item{intraclass.cor}{The intraclass correlation}
}
\description{
Computes the confidence interval for a proportion arising from
clustered data. That is, individual responses are nested within a
cluster; for example disease prevalence may be estimated on the basis
of responses given from individuals in different communities. The
clustering of responses is taken into account in the estimation of
the standard error of the proportion using the method described in
Bennet et al. (1991).
}
\details{
When the intraclass correlation is estimated to be lower
    than 0, the value is set to 0 as intraclass correlations of less
    than zero are generally implausible (Donner & Klar, 1993)
}
\examples{

# Example from Bennet et al. (1991)
ci.one.prop.cluster(95, successes = c(2, 5, 3, 3, 1, 0), N = c(2, 7, 4, 6, 4, 3))


}
\references{
Bennett, S., Woods, T., Liyanage, W. M., Smith, D. L. (1991). A
    simplified general method for cluster-sample surveys of health in
    developing countries. World Health Stat Quarterly, 44(3), 98-106.

Donner, A., & Klar, N. (1993). Confidence interval construction for
  effect measures arising from cluster randomization trials. Journal
  of clinical epidemiology, 46(2), 123-131.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
