% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_one_prop.R
\name{ci.one.prop}
\alias{ci.one.prop}
\title{Confidence interval for a single proportion}
\usage{
ci.one.prop(ci, r, n)
}
\arguments{
\item{ci}{The confidence level - pass as a natural number (e.g. 95 for the
95\% confidence interval)}

\item{r}{Frequency of events}

\item{n}{Total trial frequency}
}
\value{
A \code{list} containing the confidence interval boudaries.
  \item{p}{The proportion r/n}
  \item{l}{The lower bound of the confidence interval of p=r/n}
  \item{u}{The upper bound of the confidence interval of p=r/n}
}
\description{
This function returns the confidence interval for a single proportion using
»Wilson's score« (1927) as recommended in Newcombe (Method 3, 1998).
}
\examples{
# Examples from Newcombe (1998):
ci.one.prop(ci=95, r=81, n=263)
ci.one.prop(ci=95, r=15, n=148)

}
\references{
Newcombe, R. G. (1998). Two‐sided confidence intervals for the single
      proportion: comparison of seven methods. Statistics in medicine, 17(8),
      857-872.

  Wilson, E. B. (1927). Probable inference, the law of succession, and statistical
      inference. Journal of the American Statistical Association, 22(158), 209-212.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
