% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mixed_between.R
\name{ci.two.clustered.props}
\alias{ci.two.clustered.props}
\title{Compute the a confidence interval for the difference between two
proportions based on clustered data}
\usage{
ci.two.clustered.props(ci, successes1, successes2, N1, N2)
}
\arguments{
\item{ci}{The confidence level - pass as a natural number (e.g. 95
for the 95\% confidence interval).}

\item{successes1}{A numeric vector indicating the number of
"successes" per cluster in condition 1}

\item{successes2}{A numeric vector indicating the number of
"successes" per cluster in condition 2}

\item{N1}{A numeric vector indicating the number of observations per
cluster in condition 1}

\item{N2}{A numeric vector indicating the number of observations per
cluster in condition 2}
}
\description{
This function computes the same confidence interval as
`ci.mixed.between` but uses different input parameters.
}
\references{
Donner, A., & Klar, N. (1993). Confidence interval construction for
  effect measures arising from cluster randomization
  trials. Journal of clinical epidemiology, 46(2), 123-131.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
