% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_two_dep_props.R
\name{ci.two.dep.props}
\alias{ci.two.dep.props}
\title{Confidence interval for the difference of two dependent proportions}
\usage{
ci.two.dep.props(ci, e, f, g, h)
}
\arguments{
\item{ci}{What confidence interval - give as a natural number (e.g. 95 for the
95\% confidence interval).}

\item{e}{Number of cases for which both events are 1.}

\item{f}{Number of cases for which the first event is 1, and the second event is 0.}

\item{g}{Number of cases for which the first event is 0, and the second event is 1.}

\item{h}{Number of cases for which both events are 0.}
}
\value{
A \code{list} containing the confidence interval boudaries
  \item{d}{The difference between the two dependent proportions (d = (f-g)/n)}
  \item{l}{The lower bound of the confidence interval of d = (f-g)/n)}
  \item{u}{The upper bound of the confidence interval of d = (f-g)/n)}
}
\description{
This function returns the confidence interval of the difference of two dependent
proportions. It can be used for statistical inference, i.e. we assume that two
proportions are significantly different from each other if the confidence interval
does not contain 0. See Newcombe (Method 10; 1998).
}
\examples{
# Examples from Newcombe (1998):
ci.two.dep.props(ci=95, e=36, f=12, g=2, h=0)
ci.two.dep.props(ci=95, e=18, f=12, g=2, h=18)
ci.two.dep.props(ci=95, e=53, f=0, g=0, h=1)
ci.two.dep.props(ci=95, e=0, f=30, g=0, h=0)

# Examples from Newcombe (2001):
ci.two.dep.props(ci=95, 4, 1, 2, 25)
ci.two.dep.props(ci=95, 2, 12, 3, 17)

}
\references{
Newcombe, R. G. (1998). Improved confidence intervals for the difference 
      between binomial proportions based on paired data. Statistics in 
      medicine, 17(22), 2635-2650.

  Newcombe, R. G. (2001). Estimating the difference between differences: measurement
      of additive scale interaction for proportions. Statistics in medicine, 20(19),
      2885-2893.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
