% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_two_indep_props.R
\name{ci.two.indep.props}
\alias{ci.two.indep.props}
\title{Confidence interval for the difference of two independent proportions}
\usage{
ci.two.indep.props(ci = NULL, a = NULL, m = NULL, b = NULL, n = NULL,
  p1 = NULL, l1 = NULL, u1 = NULL, p2 = NULL, l2 = NULL, u2 = NULL)
}
\arguments{
\item{ci}{What confidence interval - give as a natural number (e.g. 95 for the
95\% confidence interval).  Only pass this argument when passing absolute
frequencies (a, m , b, n); do not combine with proportions and confidence
intervall boundaries.}

\item{a}{Frequency of events in proportion A}

\item{m}{Total frequency in proportion A}

\item{b}{Frequency of events in proportion B}

\item{n}{Total frequency in proportion B}

\item{p1}{Proportion 1; pass only if absolute frequencies (a, m , b, n) are not
given}

\item{l1}{The lower bound of the confidence interval of p1.}

\item{u1}{The lower bound of the confidence interval of p1.}

\item{p2}{Proportion 2; pass only if absolute frequencies (a, m , b, n) are not
given}

\item{l2}{The lower bound of the confidence interval of p2.}

\item{u2}{The lower bound of the confidence interval of p2.}
}
\value{
A \code{list} containing the confidence interval boudaries
  \item{d}{The difference between the proportions (a/m) - (b/n)}
  \item{l}{The lower bound of the confidence interval of d = (a/m) - (b/n)}
  \item{u}{The upper bound of the confidence interval of d = (a/m) - (b/n)}
}
\description{
This function returns the confidence interval of the difference of two independent
proportions. It can be used for statistical inference, i.e. we assume that two
proportions are significantly different from each other if the confidence interval
does not contain 0. See Newcombe (Method 10, 1998).
}
\examples{
# Examples from Newcombe (1998):
ci.two.indep.props(ci=95, a=56, m=70, b=48, n=80)
ci.two.indep.props(ci=95, a=9,  m=10, b=3,  n=10)

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
\references{
Newcombe, R. G. (1998). Interval estimation for the difference between
      independent proportions: comparison of eleven methods. Statistics in medicine,
      17(8), 873-890.
}

