% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propint-package.R
\docType{package}
\name{propint-package}
\alias{propint-package}
\title{The propint Package}
\description{
Interactions in proportions.
}
\details{
\tabular{ll}{
Package: \tab propint\cr
Type: \tab Package\cr
Version: \tab 0.01-1\cr
Date: \tab 2016-10-25\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

Compute confidence intervals for interactions in proportions
}
\author{
Martin Papenberg
}
\keyword{package}

