% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propint-package.R
\docType{package}
\name{propint-package}
\alias{propint-package}
\title{The propint Package}
\description{
Testing for interactions in proportions.
}
\details{
\tabular{ll}{
Package: \tab propint\cr
Type: \tab Package\cr
Version: \tab 0.2.13\cr
Date: \tab 2018-01-29\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL-3 + file LICENSE\cr
LazyLoad: \tab yes\cr
URL: \tab https://github.com/m-Py/propint\cr
}

propint computes confidence intervals of proportions, of their differences, and of
interactions in proportions.
}
\author{
Martin Papenberg <martin.papenberg@hhu.de>
}
\keyword{package}
